requirements.txt
langchain
pypdf
chromadb
langchain-google-genai
streamlit
python-dotenv
